Create TestSuiteCreate
This is a Powershell script to request a creation of a TestSuite using the azure RestApi. 
All the documentation of Azure, you can find here:https://learn.microsoft.com/en-us/rest/api/azure/devops/test/test-suites/create?view=azure-devops-rest-5.0&tabs=HTTP
The only way that worked for me was with 5.0 version with a Basic Auth using my personal PAT, not a System.AccessToken.